requirements.txt
streamlit>=1.30.0
pdfplumber>=0.10.0
openpyxl>=3.1.0
pandas>=2.0.0
tabula-py>=2.9.0
